Rewrote most of the code to allow for options
You can now set the number of cells to sample, whether you will sample blanks, and what counts as a blank. You can also specify variables to exclude, and the ID is automatically excluded. It also uses the data in memory instead of using a new dataset, and replaces it with the sampled cells.